feat(project): Code that requires torch is put under .torch modules
