+++ examples/demo_agent_stats.py
@@ -1,7 +1,7 @@
 
 from rlberry.envs.benchmarks.ball_exploration import PBall2D
 from rlberry.agents import RSKernelUCBVIAgent, RSUCBVIAgent
-from rlberry.agents.ppo import PPOAgent
+from rlberry.agents.torch.ppo import PPOAgent
 from rlberry.stats import AgentStats, plot_episode_rewards, compare_policies
