+++ examples/demo_agent_stats_pickle.py
@@ -1,5 +1,5 @@
 from rlberry.envs.benchmarks.ball_exploration import PBall2D
-from rlberry.agents.ppo import PPOAgent
+from rlberry.agents.torch.ppo import PPOAgent
 from rlberry.stats import AgentStats, plot_episode_rewards, compare_policies
