+++ examples/demo_hyperparam_optim.py
@@ -1,5 +1,5 @@
 from rlberry.envs.benchmarks.ball_exploration import PBall2D
-from rlberry.agents.ppo import PPOAgent
+from rlberry.agents.torch.ppo import PPOAgent
 from rlberry.stats import AgentStats
