+++ examples/demo_ppo_benchmark.py
@@ -1,6 +1,6 @@
 from rlberry.envs.benchmarks.ball_exploration.ball2d import get_benchmark_env
 from rlberry.agents import MBQVIAgent
-from rlberry.agents.ppo import PPOAgent
+from rlberry.agents.torch.ppo import PPOAgent
 from rlberry.wrappers import DiscretizeStateWrapper
 from rlberry.stats import AgentStats, plot_episode_rewards, compare_policies
