+++ examples/demo_ppo_partial_fit.py
@@ -1,5 +1,5 @@
 
-from rlberry.agents.ppo import PPOAgent
+from rlberry.agents.torch.ppo import PPOAgent
 from rlberry.envs.benchmarks.ball_exploration import PBall2D
 from rlberry.stats import AgentStats, plot_episode_rewards, compare_policies
