+++ examples/demo_rnd.py
@@ -1,4 +1,4 @@
-from rlberry.exploration_tools.rnd import RandomNetworkDistillation
+from rlberry.exploration_tools.torch.rnd import RandomNetworkDistillation
 from rlberry.envs.benchmarks.ball_exploration.ball2d import get_benchmark_env
 
 # Environment
