+++ rlberry/agents/__init__.py
@@ -1,15 +1,13 @@
-# interfaces
+# Import interfaces
 from .agent import Agent
 from .incremental_agent import IncrementalAgent
-# all agents (alphabetical!)
-from .a2c import *
+
+# basic agents (alphabetical!)
+# basic = does not require torch, jax etc
 from .adaptiveql import *
-from .avec import *
-from .dqn import *
 from .dynprog import *
 from .kernel_based import *
 from .mbqvi import *
 from .optql import *
-from .ppo import *
-from .reinforce import *
 from .ucbvi import *
+
