+++ rlberry/agents/torch/__init__.py
@@ -1,0 +1,7 @@
+# agents (alphabetical!)
+from .a2c import *
+from .avec import *
+from .dqn import *
+from .ppo import *
+from .reinforce import *
+
