+++ rlberry/agents/torch/tests/test_reinforce.py
@@ -1,5 +1,4 @@
-from copy import deepcopy
-from rlberry.agents import REINFORCEAgent
+from rlberry.agents.torch import REINFORCEAgent
 from rlberry.envs.benchmarks.ball_exploration.ball2d import get_benchmark_env
 from rlberry.exploration_tools.discrete_counter import DiscreteCounter
 from rlberry.wrappers.uncertainty_estimator_wrapper import \
