+++ rlberry/agents/torch/tests/test_torch_training.py
@@ -1,7 +1,7 @@
-from rlberry.agents.utils.torch_training import *
+from rlberry.agents.torch.utils.training import loss_function_factory, optimizer_factory
 from torch.nn import functional as F
 from rlberry.envs.benchmarks.ball_exploration.ball2d import get_benchmark_env
-from rlberry.agents.utils.torch_models import default_policy_net_fn
+from rlberry.agents.torch.utils.models import default_policy_net_fn
 
 # loss_function_factory
 assert isinstance(loss_function_factory("l2").__name__, type(F.mse_loss.__name__))
