+++ rlberry/experiment/tests/test_experiment_generator.py
@@ -2,7 +2,6 @@ from rlberry.experiment import experiment_generator
 from rlberry.agents.kernel_based.rs_ucbvi import RSUCBVIAgent
 
 import numpy as np
-import rlberry.seeding as sd
 
 
 def test_mock_args(monkeypatch):
